move common integration tests and remove duplicates
